base/pyproject.toml
[tool.poetry]
name = "base"
version = "0.1.0"
description = ""
authors = ["vlad-terehoff <tereh_off95@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
python-keycloak = "^4.2.2"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
